use a lockfile for versions

+++ updater.js
@@ -19,8 +19,26 @@ if (semver.validRange(desiredVersion) == null) {
     process.exit(1);
 }
 
+const lockfile = path.join(__dirname, 'bpac-barcode.lock');
 const outputPath = path.join(__dirname, '/bpac-barcode');
 const outputPathTemp = path.join(__dirname, `/bpac-barcode_${uuidV4()}`);
+
+let isLockfileOutdated = true;
+let lockfileContents = { desired: null, actual: null };
+if (fs.existsSync(lockfile)) {
+    lockfileContents = fs.readJsonSync(lockfile, { throws: false });
+
+    if (!lockfileContents) {
+        console.error('Lockfile invalid, cleaning up');
+        fs.remove(lockfile);
+    } else {
+        isLockfileOutdated = lockfileContents.desired !== desiredVersion;
+
+        if (isLockfileOutdated) {
+            console.warn('Lockfile is outdated');
+        }
+    }
+}
 
 bpacRepo.listReleases((error, data) => {
     if (!error) {
@@ -46,18 +64,28 @@ bpacRepo.listReleases((error, data) => {
             }
         }
 
-        if (release && asset) {
-            const downloadUrl = asset.browser_download_url;
-            download(downloadUrl, outputPathTemp, {
-                extract: true,
-            }).then(() => {
-                console.log(`Downloaded and extracted bpac-barcode:${release.tag_name}.`);
-                fs.removeSync(outputPath);
-                fs.moveSync(outputPathTemp, outputPath);
-                console.log('Replaced old with new version');
-            });
+        if (isLockfileOutdated || release.tag_name !== lockfileContents.actual) {
+            if (release && asset) {
+                const downloadUrl = asset.browser_download_url;
+                download(downloadUrl, outputPathTemp, {
+                    extract: true,
+                }).then(() => {
+                    console.info(`Downloaded and extracted bpac-barcode:${release.tag_name}.`);
+                    fs.removeSync(outputPath);
+                    fs.moveSync(outputPathTemp, outputPath);
+                    console.info('Replaced old with new version');
+                    fs.writeJson(lockfile, {
+                        desired: desiredVersion,
+                        actual: release.tag_name,
+                    }, 'utf8', () => {
+                        console.info('Wrote lock file');
+                    });
+                });
+            } else {
+                console.error(`No matching release found for ${desiredVersion}`);
+            }
         } else {
-            console.error(`No matching release found for ${desiredVersion}`);
+            console.info('bpac-barcode is up to date');
         }
     }
 });
